Change AdFailedToLoad mediatorErrorCode type from Long to Int (#2924)
## Problem
Ad networks like AdMob provide error codes as `Int`, but the
`AdFailedToLoad` event's `mediatorErrorCode` parameter expects a `Long`.
This creates friction for developers who need to manually convert the
error code type when tracking ad load failures:

```kotlin
// Before - developers had to manually convert
val loadAdError: LoadAdError = // from AdMob
adTracker.trackAdFailedToLoad(
    AdFailedToLoadData(
        // ...
        mediatorErrorCode = loadAdError.code.toLong() // Manual conversion required
    )
)
```

## Solution
Changed `mediatorErrorCode` type from `Long?` to `Int?` throughout the
SDK to match the type that ad networks naturally provide.

```kotlin
// After - developers can pass the value directly
val loadAdError: LoadAdError = // from AdMob
adTracker.trackAdFailedToLoad(
    AdFailedToLoadData(
        // ...
        mediatorErrorCode = loadAdError.code // Direct usage, no conversion needed
    )
)
```

## Changes
- `AdFailedToLoadData.mediatorErrorCode`: `Long?` → `Int?`
- `AdEvent.FailedToLoad.mediatorErrorCode`: `Long?` → `Int?`
- `BackendEvent.Ad.mediatorErrorCode`: `Long?` → `Int?`
- Updated all related unit tests to use `Int` values

## Testing
- All existing unit tests updated to reflect the type change
- No behavioral changes, only type alignment with ad network APIs

## Impact
- **Breaking change:** Developers currently passing `Long` values will
need to remove the `.toLong()` conversion
- Improves developer experience by eliminating unnecessary type
conversions
- Better alignment with ad network APIs (AdMob, AppLovin, etc.)

+++ purchases/src/main/kotlin/com/revenuecat/purchases/ads/events/types/AdFailedToLoadData.kt
@@ -19,5 +19,5 @@ class AdFailedToLoadData(
     val mediatorName: AdMediatorName,
     val placement: String?,
     val adUnitId: String,
-    val mediatorErrorCode: Long?,
+    val mediatorErrorCode: Int?,
 )
